+++ apps/web/src/app/core/state/state.service.ts
@@ -14,9 +14,9 @@ import {
 } from "@bitwarden/common/abstractions/storage.service";
 import { CollectionData } from "@bitwarden/common/admin-console/models/data/collection.data";
 import { StateFactory } from "@bitwarden/common/factories/stateFactory";
-import { SendData } from "@bitwarden/common/models/data/send.data";
 import { StorageOptions } from "@bitwarden/common/models/domain/storage-options";
 import { StateService as BaseStateService } from "@bitwarden/common/services/state.service";
+import { SendData } from "@bitwarden/common/tools/send/models/data/send.data";
 import { CipherData } from "@bitwarden/common/vault/models/data/cipher.data";
 import { FolderData } from "@bitwarden/common/vault/models/data/folder.data";
