+++ apps/web/src/app/shared/loose-components.module.ts
@@ -73,10 +73,6 @@ import { FrontendLayoutComponent } from "../layouts/frontend-layout.component";
 import { NavbarComponent } from "../layouts/navbar.component";
 import { ProductSwitcherModule } from "../layouts/product-switcher/product-switcher.module";
 import { UserLayoutComponent } from "../layouts/user-layout.component";
-import { AccessComponent } from "../send/access.component";
-import { AddEditComponent as SendAddEditComponent } from "../send/add-edit.component";
-import { EffluxDatesComponent as SendEffluxDatesComponent } from "../send/efflux-dates.component";
-import { SendComponent } from "../send/send.component";
 import { AccountComponent } from "../settings/account.component";
 import { AdjustStorageComponent } from "../settings/adjust-storage.component";
 import { ApiKeyComponent } from "../settings/api-key.component";
@@ -99,6 +95,10 @@ import { UpdateLicenseComponent } from "../settings/update-license.component";
 import { VaultTimeoutInputComponent } from "../settings/vault-timeout-input.component";
 import { GeneratorComponent } from "../tools/generator.component";
 import { PasswordGeneratorHistoryComponent } from "../tools/password-generator-history.component";
+import { AccessComponent } from "../tools/send/access.component";
+import { AddEditComponent as SendAddEditComponent } from "../tools/send/add-edit.component";
+import { EffluxDatesComponent as SendEffluxDatesComponent } from "../tools/send/efflux-dates.component";
+import { SendComponent } from "../tools/send/send.component";
 import { ToolsComponent } from "../tools/tools.component";
 import { PasswordRepromptComponent } from "../vault/components/password-reprompt.component";
 import { PremiumBadgeComponent } from "../vault/components/premium-badge.component";
