+++ apps/web/src/app/tools/send/send.component.ts
@@ -1,17 +1,17 @@
 import { Component, NgZone, ViewChild, ViewContainerRef } from "@angular/core";
 
-import { SendComponent as BaseSendComponent } from "@bitwarden/angular/components/send/send.component";
 import { ModalService } from "@bitwarden/angular/services/modal.service";
+import { SendComponent as BaseSendComponent } from "@bitwarden/angular/tools/send/send.component";
 import { BroadcasterService } from "@bitwarden/common/abstractions/broadcaster.service";
 import { EnvironmentService } from "@bitwarden/common/abstractions/environment.service";
 import { I18nService } from "@bitwarden/common/abstractions/i18n.service";
 import { LogService } from "@bitwarden/common/abstractions/log.service";
 import { PlatformUtilsService } from "@bitwarden/common/abstractions/platformUtils.service";
 import { SearchService } from "@bitwarden/common/abstractions/search.service";
-import { SendApiService } from "@bitwarden/common/abstractions/send/send-api.service.abstraction";
-import { SendService } from "@bitwarden/common/abstractions/send/send.service.abstraction";
 import { PolicyService } from "@bitwarden/common/admin-console/abstractions/policy/policy.service.abstraction";
-import { SendView } from "@bitwarden/common/models/view/send.view";
+import { SendView } from "@bitwarden/common/tools/send/models/view/send.view";
+import { SendApiService } from "@bitwarden/common/tools/send/services/send-api.service.abstraction";
+import { SendService } from "@bitwarden/common/tools/send/services/send.service.abstraction";
 import { Icons } from "@bitwarden/components";
 
 import { AddEditComponent } from "./add-edit.component";
